Added missing rounding when calculating the FIFO costs

+++ name.abuchen.portfolio/src/name/abuchen/portfolio/snapshot/security/CostCalculation.java
@@ -64,8 +64,8 @@ import name.abuchen.portfolio.money.Money;
 
                     long n = Math.min(sold, entry.shares);
 
-                    entry.grossAmount -= Math.round(n * entry.grossAmount / entry.shares);
-                    entry.netAmount -= Math.round(n * entry.netAmount / entry.shares);
+                    entry.grossAmount -= Math.round(n * entry.grossAmount / (double) entry.shares);
+                    entry.netAmount -= Math.round(n * entry.netAmount / (double) entry.shares);
                     entry.shares -= n;
 
                     sold -= n;
